+++ service/indexer/internal/server/server.go
@@ -36,10 +36,12 @@ import (
 	"github.com/naturalselectionlabs/pregod/service/indexer/internal/datasource/conflux"
 	"github.com/naturalselectionlabs/pregod/service/indexer/internal/datasource/eip1577"
 	"github.com/naturalselectionlabs/pregod/service/indexer/internal/datasource/kurora"
+
+	"github.com/naturalselectionlabs/pregod/service/indexer/internal/datasource/kurora/ethereum"
 	"github.com/naturalselectionlabs/pregod/service/indexer/internal/datasource/kurora/lens"
 	"github.com/naturalselectionlabs/pregod/service/indexer/internal/datasource/kurora/mirror"
 	"github.com/naturalselectionlabs/pregod/service/indexer/internal/datasource/moralis"
-	eth_etl "github.com/naturalselectionlabs/pregod/service/indexer/internal/datasource/pregod_etl/ethereum"
+
 	"github.com/naturalselectionlabs/pregod/service/indexer/internal/datasource/zksync"
 	"github.com/naturalselectionlabs/pregod/service/indexer/internal/datasource_asset"
 	"github.com/naturalselectionlabs/pregod/service/indexer/internal/datasource_asset/nftscan"
@@ -174,6 +176,11 @@ func (s *Server) Initialize() (err error) {
 		return err
 	}
 
+	ethereumDatasource, err := ethereum.New(context.Background(), s.config.Kurora.Endpoint)
+	if err != nil {
+		return err
+	}
+
 	s.datasources = []datasource.Datasource{
 		kuroraDatasource,
 		mirrorDatasource,
@@ -181,7 +188,7 @@ func (s *Server) Initialize() (err error) {
 		moralis.New(s.config.Moralis.Key),
 		blockscoutDatasource,
 		zksync.New(),
-		eth_etl.New(),
+		ethereumDatasource,
 		eip1577.New(s.employer),
 		lensDatasource,
 		aptos.New(),
